web/requirements.txt
Django==2.2.10
django-allauth==0.41.0
django-ckeditor==5.9.0
django-crispy-forms==1.9.0
django-haystack==2.4.1
django-js-asset==1.2.2
django-messages==0.6.0
django-nocaptcha-recaptcha==0.0.20
django-pagination-py3==1.1.2
django-widget-tweaks==1.4.5
djangorestframework
Faker==4.0.1
gunicorn==20.0.4
Markdown==3.2.1
markdown2==2.3.8
Pillow==7.0.0
postmarkup==1.2.2
psycopg2-binary==2.8.4
Pygments==2.5.2
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.0
text-unidecode==1.3
urllib3==1.25.8
Whoosh==2.7.4
